Move monads to separated package `mona.monads`
