+++ mona/res/body.py
@@ -3,7 +3,8 @@ import typing
 import orjson
 import pydantic
 
-from mona import context, future, handler, state
+from mona import context, handler
+from mona.monads import future, state
 
 from . import header, status
