+++ mona/res/header.py
@@ -1,4 +1,5 @@
-from mona import context, handler, state
+from mona import context, handler
+from mona.monads import state
 
 
 def set_header(key: str, value: str) -> handler.Handler:
