+++ tests/test_request_type.py
@@ -1,6 +1,7 @@
 import pytest
 
-from mona import context, req, state
+from mona import context, req
+from mona.monads import state
 
 
 @pytest.mark.parametrize(
