+++ tests/test_response_header.py
@@ -1,6 +1,7 @@
 import pytest
 
-from mona import context, future, res, state
+from mona import context, res
+from mona.monads import future, state
 
 
 @pytest.mark.asyncio
